feat(user-ui): update TailwindCSS CLI and refactor main page components
+ Replace deprecated TailwindCSS CLI with the latest version
+ Refactor and reorganize main page components for better modularity
+ Update package-lock.json to include new dependencies and remove unused ones
