+++ packages/user-ui/package.json
@@ -9,6 +9,7 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "@tailwindcss/cli": "^4.0.12",
     "firebase": "^11.3.1"
   },
   "devDependencies": {
