--- packages/user-ui/src/pages/main/mod.rs
@@ -1,7 +1,0 @@
-mod controller;
-mod i18n;
-
-pub mod page;
-pub use page::MainPage;
-
-pub mod components;
