+++ packages/user-ui/src/pages/mod.rs
@@ -1,8 +1,14 @@
+mod controller;
 pub mod governance;
-pub mod main;
+mod i18n;
 pub mod not_found;
 pub mod profile;
 pub mod project;
 pub mod users;
+
+pub mod page;
+pub use page::MainPage;
+
+mod components;
 
 pub use not_found::NotFoundPage;
